refactor(test): 컨트롤러 테스트가 아닌 통합테스트로 이름 변경
